fix: 位运算变量从int32_t改为uint32_t避免有符号位运算未定义行为
system_sound_manager_impl.cpp中RingtoneTypeToBitMask/SystemToneTypeToBitMask等
函数及相关的位掩码变量使用了有符号int32_t进行位运算(<<, &, |, ~)，
有符号整数的位运算可能导致实现定义行为或未定义行为。
将TONE_BIT_BASE常量、RingtoneTypeToBitMask/SystemToneTypeToBitMask返回类型、
UpdateToneTypeParams结构体成员、以及所有参与位运算的局部变量
(targetToneType, typeValue, oldType, remaining, finalType, storedToneType)
从int32_t改为uint32_t。
在GetRingtoneType/GetShottoneType返回值边界处添加static_cast<uint32_t>，
在DataShareValuesBucket::Put调用处添加static_cast<int32_t>保持接口兼容。

+++ frameworks/native/system_sound_manager/system_sound_manager_impl.cpp
@@ -73,7 +73,7 @@ const int32_t MAX_VECTOR_LENGTH = 1024;
 const off_t MAX_FILE_SIZE_200M = 200 * 1024 * 1024;
 const int32_t PARAM1 = 1;
 const int32_t PARAM2 = 2;
-const int32_t TONE_BIT_BASE = 1;
+const uint32_t TONE_BIT_BASE = 1;
 const int32_t RINGER_MODE_OFFSET_RING_OR_SILENT = 0;
 const int32_t RINGER_MODE_OFFSET_VIBRATE = 1;
 const int32_t TONE_SETTING_TYPE_BASE = 100;
@@ -244,18 +244,18 @@ std::string SystemSoundManagerImpl::BuildVibrateLibraryUri(bool isProxy) const
         std::to_string(SystemSoundManagerUtils::GetCurrentUserId()) : VIBRATE_PATH_URI;
 }
 
-int32_t SystemSoundManagerImpl::RingtoneTypeToBitMask(RingtoneType ringtoneType) const
+uint32_t SystemSoundManagerImpl::RingtoneTypeToBitMask(RingtoneType ringtoneType) const
 {
     return TONE_BIT_BASE << (static_cast<int32_t>(ringtoneType) - RINGTONE_TYPE_SIM_CARD_0);
 }
 
-int32_t SystemSoundManagerImpl::SystemToneTypeToBitMask(SystemToneType systemToneType) const
+uint32_t SystemSoundManagerImpl::SystemToneTypeToBitMask(SystemToneType systemToneType) const
 {
     return TONE_BIT_BASE << (static_cast<int32_t>(systemToneType) - SYSTEM_TONE_TYPE_SIM_CARD_0);
 }
 
 ToneAttrs SystemSoundManagerImpl::QueryToneAttrsByType(const DatabaseTool &databaseTool,
-    const std::string &typeColumnName, int32_t targetToneType, SourceType sourceType, int32_t defaultCategory)
+    const std::string &typeColumnName, uint32_t targetToneType, SourceType sourceType, int32_t defaultCategory)
 {
     ToneAttrs toneAttrs = { "", "", "", CUSTOMISED, defaultCategory };
     CHECK_AND_RETURN_RET_LOG(databaseTool.isInitialized && databaseTool.dataShareHelper != nullptr, toneAttrs,
@@ -284,8 +284,8 @@ ToneAttrs SystemSoundManagerImpl::QueryToneAttrsByType(const DatabaseTool &datab
 
     unique_ptr<RingtoneAsset> asset = results->GetFirstObject();
     while (asset != nullptr) {
-        int32_t typeValue = typeColumnName == RINGTONE_COLUMN_RING_TONE_TYPE ?
-            asset->GetRingtoneType() : asset->GetShottoneType();
+        uint32_t typeValue = typeColumnName == RINGTONE_COLUMN_RING_TONE_TYPE ?
+            static_cast<uint32_t>(asset->GetRingtoneType()) : static_cast<uint32_t>(asset->GetShottoneType());
         if (typeValue & targetToneType) {
             toneAttrs.SetUri(asset->GetPath());
             toneAttrs.SetTitle(asset->GetTitle());
@@ -349,7 +349,7 @@ int32_t SystemSoundManagerImpl::ClearNotificationToneType(
 int32_t SystemSoundManagerImpl::ClearBitFromToneTypeColumn(
     std::shared_ptr<DataShare::DataShareHelper> dataShareHelper,
     const std::string &typeColumnName, const std::string &sourceTypeColumnName,
-    int32_t targetToneType, SourceType sourceType)
+    uint32_t targetToneType, SourceType sourceType)
 {
     int32_t result = 0;
     DataShare::DatashareBusinessError businessError;
@@ -363,13 +363,13 @@ int32_t SystemSoundManagerImpl::ClearBitFromToneTypeColumn(
 
     unique_ptr<RingtoneAsset> asset = results->GetFirstObject();
     while (asset != nullptr) {
-        int32_t oldType;
+        uint32_t oldType;
         if (typeColumnName == RINGTONE_COLUMN_RING_TONE_TYPE) {
-            oldType = asset->GetRingtoneType();
+            oldType = static_cast<uint32_t>(asset->GetRingtoneType());
         } else {
-            oldType = asset->GetShottoneType();
+            oldType = static_cast<uint32_t>(asset->GetShottoneType());
         }
-        int32_t remaining = oldType & ~targetToneType;
+        uint32_t remaining = oldType & ~targetToneType;
 
         DataSharePredicates updatePredicates;
         DataShareValuesBucket updateValuesBucket;
@@ -393,7 +393,7 @@ int32_t SystemSoundManagerImpl::ClearBitFromToneTypeColumn(
                 result += dataShareHelper->Update(RINGTONEURI, updatePredicates, updateValuesBucket);
             }
         } else {
-            updateValuesBucket.Put(typeColumnName, remaining);
+            updateValuesBucket.Put(typeColumnName, static_cast<int32_t>(remaining));
             result += dataShareHelper->Update(RINGTONEURI, updatePredicates, updateValuesBucket);
         }
         asset = results->GetNextObject();
@@ -460,18 +460,18 @@ int32_t SystemSoundManagerImpl::UpdateToneTypeUri(std::shared_ptr<DataShare::Dat
         params.targetToneType, SOURCE_TYPE_CUSTOMISED);
 
     // Step 2: Set new target type on the target tone
-    int32_t finalType = params.targetToneType | params.storedToneType;
+    uint32_t finalType = params.targetToneType | params.storedToneType;
     DataSharePredicates updatePredicates;
     DataShareValuesBucket updateValuesBucket;
     updatePredicates.SetWhereClause(RINGTONE_COLUMN_TONE_ID + " = ? ");
     updatePredicates.SetWhereArgs({to_string(params.toneId)});
-    updateValuesBucket.Put(params.typeColumnName, finalType);
+    updateValuesBucket.Put(params.typeColumnName, static_cast<int32_t>(finalType));
     updateValuesBucket.Put(params.sourceTypeColumnName, SOURCE_TYPE_CUSTOMISED);
     return dataShareHelper->Update(RINGTONEURI, updatePredicates, updateValuesBucket);
 }
 
 int32_t SystemSoundManagerImpl::UpdateRingtoneUri(std::shared_ptr<DataShare::DataShareHelper> dataShareHelper,
-    const int32_t &toneId, RingtoneType ringtoneType, const int32_t &storedToneType)
+    const int32_t &toneId, RingtoneType ringtoneType, const uint32_t &storedToneType)
 {
     UpdateToneTypeParams params = {
         toneId,
@@ -498,8 +498,9 @@ int32_t SystemSoundManagerImpl::SetToneUriInternal(std::shared_ptr<DataShare::Da
 
     unique_ptr<RingtoneAsset> ringtoneAsset = results->GetFirstObject();
     if (ringtoneAsset != nullptr) {
-        int32_t storedToneType = (params.toneTypeQuery == TONE_TYPE_RINGTONE) ?
-            ringtoneAsset->GetRingtoneType() : ringtoneAsset->GetShottoneType();
+        uint32_t storedToneType = (params.toneTypeQuery == TONE_TYPE_RINGTONE) ?
+            static_cast<uint32_t>(ringtoneAsset->GetRingtoneType()) :
+            static_cast<uint32_t>(ringtoneAsset->GetShottoneType());
 
         int32_t toneId = ringtoneAsset->GetId();
         int32_t changedRows = 0;
@@ -547,7 +548,7 @@ int32_t SystemSoundManagerImpl::SetNoRingToneUri(std::shared_ptr<DataShare::Data
 {
     MEDIA_LOGI("Set no audio uri for ringtone type %{public}d", ringtoneType);
     int32_t result = 0;
-    int32_t targetToneType = RingtoneTypeToBitMask(ringtoneType);
+    uint32_t targetToneType = RingtoneTypeToBitMask(ringtoneType);
     result += ClearBitFromToneTypeColumn(dataShareHelper, RINGTONE_COLUMN_RING_TONE_TYPE,
         RINGTONE_COLUMN_RING_TONE_SOURCE_TYPE, targetToneType, SOURCE_TYPE_CUSTOMISED);
     result += ClearBitFromToneTypeColumn(dataShareHelper, RINGTONE_COLUMN_RING_TONE_TYPE,
@@ -661,7 +662,7 @@ ToneAttrs SystemSoundManagerImpl::GetRingtoneAttrs(const DatabaseTool &databaseT
         return toneAttrs;
     }
 
-    int32_t targetToneType = RingtoneTypeToBitMask(ringtoneType);
+    uint32_t targetToneType = RingtoneTypeToBitMask(ringtoneType);
     toneAttrs = QueryToneAttrsByType(databaseTool, RINGTONE_COLUMN_RING_TONE_TYPE,
         targetToneType, SOURCE_TYPE_CUSTOMISED, TONE_CATEGORY_RINGTONE);
     if (toneAttrs.GetUri().empty()) {
@@ -752,7 +753,7 @@ std::shared_ptr<SystemTonePlayer> SystemSoundManagerImpl::GetSystemTonePlayer(
 }
 
 int32_t SystemSoundManagerImpl::UpdateShotToneUri(std::shared_ptr<DataShare::DataShareHelper> dataShareHelper,
-    const int32_t &toneId, SystemToneType systemToneType, const int32_t &storedToneType)
+    const int32_t &toneId, SystemToneType systemToneType, const uint32_t &storedToneType)
 {
     UpdateToneTypeParams params = {
         toneId,
@@ -787,7 +788,7 @@ int32_t SystemSoundManagerImpl::SetNoSystemToneUri(std::shared_ptr<DataShare::Da
         result += ClearNotificationToneType(dataShareHelper, SOURCE_TYPE_CUSTOMISED);
         result += ClearNotificationToneType(dataShareHelper, SOURCE_TYPE_PRESET);
     } else {
-        int32_t targetToneType = SystemToneTypeToBitMask(systemToneType);
+        uint32_t targetToneType = SystemToneTypeToBitMask(systemToneType);
         result += ClearBitFromToneTypeColumn(dataShareHelper, RINGTONE_COLUMN_SHOT_TONE_TYPE,
             RINGTONE_COLUMN_SHOT_TONE_SOURCE_TYPE, targetToneType, SOURCE_TYPE_CUSTOMISED);
         result += ClearBitFromToneTypeColumn(dataShareHelper, RINGTONE_COLUMN_SHOT_TONE_TYPE,
@@ -866,7 +867,7 @@ ToneAttrs SystemSoundManagerImpl::GetSystemToneAttrs(const DatabaseTool &databas
                 to_string(NOTIFICATION_TONE_TYPE), SOURCE_TYPE_PRESET);
         }
     } else {
-        int32_t targetToneType = SystemToneTypeToBitMask(systemToneType);
+        uint32_t targetToneType = SystemToneTypeToBitMask(systemToneType);
         toneAttrs = QueryToneAttrsByType(databaseTool, RINGTONE_COLUMN_SHOT_TONE_TYPE,
             targetToneType, SOURCE_TYPE_CUSTOMISED, category);
         if (toneAttrs.GetUri().empty()) {
